feat: two bug fixed user_id type changed to BIGINT application.yml url updated

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE users(
-                      user_id INT AUTO_INCREMENT,
+                      user_id BIGINT AUTO_INCREMENT,
                       first_name VARCHAR(50) NOT NULL,
                       last_name VARCHAR(50) NOT NULL ,
                       national_id CHAR(10) NOT NULL UNIQUE ,
